# docs/swagger.yaml
basePath: /v1
definitions:
...
  main.RegisterExperiencePayload:
    properties:
      company:
        type: string
      description:
        type: string
      title:
        type: string
      year_from:
        type: string
      year_to:
        type: string
    required:
    - title
    - year_from
    type: object
...
  main.UpdateExperiencePayload:
    properties:
      company:
        type: string
      description:
        type: string
      title:
        type: string
      year_from:
        type: string
      year_to:
        type: string
    type: object
...
  store.Experience:
    properties:
      company:
        type: string
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      userid:
        type: integer
      year_from:
        type: string
      year_to:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a Appointr server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Appointr API
  version: "1.0"
paths:
...
  /experience/{experienceID}:
    delete:
      consumes:
      - application/json
      description: Delete experience by id
      parameters:
      - description: Experience ID
        in: path
        name: experienceID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Delete experience
      tags:
      - experience
    patch:
      consumes:
      - application/json
      description: Update experience
      parameters:
      - description: Experience ID
        in: path
        name: experienceID
        required: true
        type: integer
      - description: Experience
        in: body
        name: experience
        required: true
        schema:
          $ref: '#/definitions/main.UpdateExperiencePayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Experience'
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Update experience
      tags:
      - experience
  /experience/{experienceid}:
    get:
      consumes:
      - application/json
      description: Get experience by ID
      parameters:
      - description: Experience ID
        in: path
        name: experienceid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Experience'
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Get experience by ID
      tags:
      - experience
  /experience/create:
    post:
      consumes:
      - application/json
      description: Create a new experience
      parameters:
      - description: Experience
        in: body
        name: experience
        required: true
        schema:
          $ref: '#/definitions/main.RegisterExperiencePayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Create a new experience
      tags:
      - experience
  /experience/u/{id}:
    get:
      consumes:
      - application/json
      description: Get all experience by user ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.Experience'
            type: array
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Get all experience by user ID
      tags:
      - experience
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"